Added error handle/ retrofit work!

+++ app/src/main/java/com/example/onlineshop/datamodels/items/LatestData.kt
@@ -6,6 +6,6 @@ import com.google.gson.annotations.SerializedName
 data class LatestData(
     @SerializedName("category") var category: String? = null,
     @SerializedName("name") var name: String? = null,
-    @SerializedName("price") var price: Int? = null,
+    @SerializedName("price") var price: Double? = null,
     @SerializedName("image_url") var imageUrl: String? = null
 )
